EES-7042 - Moving some services to a `UserManagement` folder to mirror the back-end controllers, and organise the proxy endpoints
