+++ src/explore-education-statistics-admin/src/pages/release/pre-release/components/PreReleaseInvitePlanModal.tsx
@@ -2,7 +2,7 @@ import styles from '@admin/pages/release/pre-release/components/PreReleaseInvite
 import ModalConfirm from '@common/components/ModalConfirm';
 import WarningMessage from '@common/components/WarningMessage';
 import React from 'react';
-import { PreReleaseInvitePlan } from '@admin/services/preReleaseUserService';
+import { PreReleaseInvitePlan } from '@admin/services/user-management/preReleaseUserService';
 
 interface Props {
   invitePlan: PreReleaseInvitePlan;
