+++ src/explore-education-statistics-admin/src/pages/users/components/__tests__/RoleForm.test.tsx
@@ -1,6 +1,6 @@
 import RoleForm from '@admin/pages/users/components/RoleForm';
 import { testUser, testRoles } from '@admin/pages/users/__data__/testUserData';
-import _userService from '@admin/services/userService';
+import _userService from '@admin/services/user-management/userService';
 import { render, screen, waitFor, within } from '@testing-library/react';
 import React from 'react';
 import noop from 'lodash/noop';
